Add entity e ajustes
Adição da configuração do entity para o modelo de produtos, ajuste das regras das rotas e da interface do swagger

+++ Compras_API_dot_net_core/Objeto/Pagamento.cs
@@ -2,7 +2,7 @@
 {
     public class Pagamento
     {
-        public float valor { get; set; }
+        public decimal valor { get; set; }
         public Cartao cartao { get; set; }
     }
 }
