+++ Compras_API_dot_net_core/Objeto/Produto.cs
@@ -3,7 +3,7 @@
     public class Produto
     {
         public string nome { get; set; }
-        public float valor_unitario { get; set; }
+        public decimal valor_unitario { get; set; }
         public int qtde_estoque { get; set; }
         
     }
